+++ doc/releaseNotes.md
@@ -648,3 +648,6 @@ This file describes the main feature changes for each readout.exe released versi
 
 ## v2.26.3 - 22/10/2024
 - Minor release for osx compatibility.
+
+## next version
+- Added symbolic links with relative path in configuration. See @LINK syntax.